requirements.txt
django==4.1.4
numpy
opencv-python
openai==0.28.1
pytesseract
python-dotenv